[IOTDB-1252] optimize test mult-times perform (#2887)
